+++ sklearnex/linear_model/tests/test_linear.py
@@ -16,22 +16,33 @@
 # ===============================================================================
 
 import numpy as np
+import pytest
 from numpy.testing import assert_allclose
 from sklearn.datasets import make_regression
 
 from daal4py.sklearn._utils import daal_check_version
+from onedal.tests.utils._dataframes_support import (
+    _as_numpy,
+    _convert_to_dataframe,
+    get_dataframes_and_queues,
+)
 
 
+@pytest.mark.parametrize("dataframe,queue", get_dataframes_and_queues())
 def test_sklearnex_import_linear():
     from sklearnex.linear_model import LinearRegression
 
     X = np.array([[1, 1], [1, 2], [2, 2], [2, 3]])
     y = np.dot(X, np.array([1, 2])) + 3
+    X = _convert_to_dataframe(X, sycl_queue=queue, target_df=dataframe)
+    y = _convert_to_dataframe(y, sycl_queue=queue, target_df=dataframe)
     linreg = LinearRegression().fit(X, y)
+    if daal_check_version((2023, "P", 100)):
+        assert hasattr(linreg, "_onedal_estimator")
     assert "sklearnex" in linreg.__module__
     assert linreg.n_features_in_ == 2
-    assert_allclose(linreg.intercept_, 3.0)
-    assert_allclose(linreg.coef_, [1.0, 2.0])
+    assert_allclose(_as_numpy(linreg.intercept_), 3.0)
+    assert_allclose(_as_numpy(linreg.coef_), [1.0, 2.0])
 
 
 def test_sklearnex_import_ridge():
